Introducing concept of a game map.
This is the first change of a major refactoring effort to associate a
map with the snake game. This will remove the need to make any
comparisons based on "top" and "left" style attributes of certain
pieces, and will allow for an O(1) comparison when determining if a
piece is on a given spot on the board.